src/service/mystic/pyproject.toml
[tool.poetry]
name = "mystic-requirements"
version = "0.1.0"
description = "mystic requirements"
authors = ["Hyunjoon Cho <juatis96@kookmin.ac.kr>"]
readme = "README.md"
packages = [
    { include = "src/app" },
]

[tool.poetry.dependencies]
torch = "*"
python = ">=3.11,<3.12"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.0.51"}
pydantic = "<2"
langchain-openai = "^0.0.8"
python-decouple = "^3.8"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.8.0"
pymongo = "^4.6.2"
pydub = "^0.25.1"
gtts = "^2.5.1"
transformers = "^4.39.3"


pillow = "^10.3.0"
tts = "^0.22.0"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
